ALL FUNCTIONS.java
push(int a): This function is used to push an element onto the stack. It takes an integer value as input and adds it to the top of the stack.

display(): This function displays all the elements of the stack starting from the top.

pop(): This function removes and returns the top element from the stack.

peep(int i1): This function allows peeking at an element at a specific position in the stack without removing it. It takes an integer parameter representing the position from the top of the stack.

change(int po, int val): This function changes the value of an element at a specified position in the stack. It takes two integer parameters: the position (po) where the value needs to be changed and the new value (val) to be assigned.

peek(): This function allows peeking at the element present at the top of the stack without removing it.

count(): This function counts and displays the total number of elements in the stack.

isEmpty(): This function checks if the stack is empty and prints a corresponding message.

isFull(): This function checks if the stack is full and prints a corresponding message.